Add at= time-travel to the market-orders and industry-jobs API endpoints
The /api/character/orders, /api/character/jobs, and /api/corp/jobs IMPORTDATA
endpoints now accept an ?at= timestamp, reconstructing the snapshot at any past
moment from the SCD-2 history — mirroring what /api/character/assets already
does. Each backing function (character_orders, character_industry_jobs,
corp_industry_jobs) gains an as_of parameter (default now) and reads its
*_over_time history table with the same validity predicate character_asset_
snapshot_at uses: keep the version of every order/job that had started by as_of
and was either still open then (valid_until >= as_of) or is the current version.
For jobs the include_delivered filter applies to the as-of version's status, so
a job that is delivered now but was active at as_of shows as active. At the
default now() each returns the same is_current set the views exposed, so
existing callers are unaffected.

The `at` parsing/validation (partial-ISO padding, default now) is factored out
of the assets route into src/utils/atParam.ts and shared by all four endpoints.

Migration drops the prior function signatures (an added defaulted param would
make the old form ambiguous) and recreates them. Validated against a throwaway
Postgres with multi-version seed data: at now() the functions return the live
is_current set; at a past timestamp they return the version valid then (an order
mid-fill, a job still active before delivery), and the migration cleanly swaps
the signatures.
