+++ src/app/api/character/assets/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server'
 
 import { resolvePlayer } from '@/utils/apiToken'
+import { AT_PARAM_ERROR, parseAtParam } from '@/utils/atParam'
 import { toCsv } from '@/utils/csv'
 
 // Public CSV endpoint for Google Sheets =IMPORTDATA(): the player's raw asset
@@ -12,33 +13,17 @@ export const dynamic = 'force-dynamic'
 // Headroom over Vercel's default function timeout for a large inventory.
 export const maxDuration = 60
 
-// Pad a partial ISO date/time out to a full UTC timestamp so callers can pass
-// just the precision they care about: 2026 → 2026-01-01T00:00:00Z, 2026-05 →
-// 2026-05-01T00:00:00Z, 2026-05-30T18 → 2026-05-30T18:00:00Z, etc. Inputs that
-// already carry fractional seconds or a timezone (e.g. a full toISOString) don't
-// match the bare-prefix pattern and are returned untouched for Date to parse.
-const completePartialAt = (value: string): string => {
-  const m = /^(\d{4})(?:-(\d{2}))?(?:-(\d{2}))?(?:[T ](\d{2}))?(?::(\d{2}))?(?::(\d{2}))?$/.exec(value)
-  if (!m) return value
-  const [, year, month = '01', day = '01', hour = '00', minute = '00', second = '00'] = m
-  return `${year}-${month}-${day}T${hour}:${minute}:${second}Z`
-}
-
 export const GET = async (request: NextRequest): Promise<NextResponse> => {
   const { searchParams } = new URL(request.url)
 
   // `at` is the moment to reconstruct the inventory at; default to now (the live
   // inventory). character_asset_over_time keeps full SCD-2 history, so any past
   // time works.
-  const atParam = searchParams.get('at')
-  const at = atParam ? new Date(completePartialAt(atParam.trim())) : new Date()
-  if (Number.isNaN(at.getTime())) {
-    return NextResponse.json(
-      { error: 'Invalid `at` timestamp; use ISO 8601 (e.g. 2026-06-01T00:00:00Z)' },
-      { status: 400 }
-    )
+  const at = parseAtParam(searchParams.get('at'))
+  if (!at.ok) {
+    return NextResponse.json({ error: AT_PARAM_ERROR }, { status: 400 })
   }
-  const atIso = at.toISOString()
+  const atIso = at.iso
 
   const player = await resolvePlayer(searchParams.get('token')?.trim())
   if (!player.ok) {
