+++ src/app/api/character/orders/route.ts
@@ -1,19 +1,27 @@
 import { NextRequest, NextResponse } from 'next/server'
 
 import { resolvePlayer } from '@/utils/apiToken'
+import { AT_PARAM_ERROR, parseAtParam } from '@/utils/atParam'
 import { toCsv } from '@/utils/csv'
 
 // Public CSV endpoint for Google Sheets =IMPORTDATA(): the player's open market
-// orders across all of their characters, with the owning character's name. The
-// first row is the column headers. Authenticated by the per-user api_token in the
-// query string (Sheets carries no session cookie), so it always recomputes — no
-// caching.
+// orders across all of their characters, with the owning character's name, as of
+// an optional `at` timestamp. The first row is the column headers. Authenticated
+// by the per-user api_token in the query string (Sheets carries no session
+// cookie), so it always recomputes — no caching.
 export const dynamic = 'force-dynamic'
 // Headroom over Vercel's default function timeout.
 export const maxDuration = 60
 
 export const GET = async (request: NextRequest): Promise<NextResponse> => {
   const { searchParams } = new URL(request.url)
+
+  // `at` reconstructs which orders were open at that moment from the SCD-2
+  // history (character_order_over_time); default now is the live open set.
+  const at = parseAtParam(searchParams.get('at'))
+  if (!at.ok) {
+    return NextResponse.json({ error: AT_PARAM_ERROR }, { status: 400 })
+  }
 
   const player = await resolvePlayer(searchParams.get('token')?.trim())
   if (!player.ok) {
@@ -24,6 +32,7 @@ export const GET = async (request: NextRequest): Promise<NextResponse> => {
   // the field order for the sheet's columns and sidesteps PostgREST's max-rows cap.
   const { data: rows, error: rowsError } = await player.supabase.rpc('character_orders', {
     character_ids: player.characterIds,
+    as_of: at.iso,
   })
   if (rowsError) {
     return NextResponse.json({ error: 'Query failed' }, { status: 500 })
